backend/pyproject.toml
[project]
name = "memoctopus-backend"
version = "0.1.0"
description = "FastAPI backend for OpenAI API passthrough"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "asyncpg>=0.29.0",
    "databases[postgresql]>=0.9.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
